+++ src/types.ts
@@ -1,45 +1,57 @@
+export type IgnoreMissingKeys = {
+    ignoreMissingKeys: boolean;
+}
+
 export type TemplateData = {
     [key:string] : string | Date
 };
 
-export type Constructor = {
+export type BrotherSDKOptions = {
     templatePath: string;
     exportDir?: string;
     printer?: string;
 };
 
-export type PrintOptions = {
+export type PrintConfig = {
     copies?: number;
     printName?: string;
 };
 
+export enum PrintOptionFlag {
+    autoCut = 0x1,
+    cutPause = 0x1,
+    cutMark = 0x2,
+    halfCut = 0x200,
+    chainPrint = 0x400,
+    tailCut = 0x800,
+    specialTape = 0x00080000,
+    cutAtEnd = 0x04000000,
+    noCut = 0x10000000,
+    mirroring = 0x4,
+    quality = 0x00010000,
+    highSpeed = 0x01000000,
+    highResolution = 0x02000000,
+    color = 0x8,
+    mono = 0x10000000,
+}
+
 export type StartPrintOptions = {
-    autoCut?: boolean;
-    cutPause?: boolean,
-    cutMark?: boolean,
-    halfCut?: boolean,
-    chainPrint?: boolean,
-    tailCut?: boolean,
-    specialTape?: boolean,
-    cutAtEnd?: boolean,
-    noCut?: boolean,
-    mirroring?: boolean,
-    quality?: boolean,
-    highSpeed?: boolean,
-    highResolution?: boolean,
-    color?: boolean,
-    mono?: boolean,
+    [key in keyof typeof PrintOptionFlag]?: boolean;
 };
 
 export type FitPage = {
     fitPage?: boolean;
 };
 
-export type PrintConfig = PrintOptions & StartPrintOptions & FitPage;
+export type PrintOptions = IgnoreMissingKeys & PrintConfig & StartPrintOptions & FitPage;
 
-export type ImageOptions = {
+export type ImageOptions = IgnoreMissingKeys & {
     width?: number;
     height?:number;
+};
+
+export type ExportOptions = IgnoreMissingKeys & {
+    resolution?: number;
 };
 
 export enum ObjectTypes {
@@ -48,4 +60,22 @@ export enum ObjectTypes {
     Image = 2,
     DateTime = 3,
     ClipArt = 4,
+}
+
+export interface PrinterStatus {
+    printerName: string | null;
+    online: boolean | null;
+    supported: boolean | null;
+    errorCode: number | null;
+    errorString: string | null;
+    currentMedia: string[] | null;
+    supportedMedia: string[] | null;
+}
+
+export enum ExportType {
+    ".lbx" = 0x1,
+    ".lbl" = 0x2,
+    ".lbi" = 0x3,
+    ".bmp" = 0x4,
+    ".paf" = 0x5,
 }