+++ tsup.config.js
@@ -4,12 +4,12 @@ const inDev = process.env.NODE_ENV !== "production";
 const inProd = process.env.NODE_ENV === "production"
 
 export default defineConfig({
-    entry: ["./src/index.ts"],
-    clean: true,
-    splitting: false,
-    sourcemap: inDev,
-    minify: inProd,
-    bundle: true,
+    entry: ["src/index.ts"],
     format: ["esm"],
     dts: true,
+    sourcemap: inDev,
+    clean: true,
+    splitting: false,
+    minify: inProd,
+    outDir: "dist",
 });